+++ tests/test_config.py
@@ -160,6 +160,34 @@ def test_get_preamble_length_bytes(transceiver, mdmcfg1, length):
     transceiver._spi.xfer.assert_called_once_with([0x13 | 0x80, 0])
 
 
+@pytest.mark.parametrize(
+    ("mdmcfg1_before", "mdmcfg1_after", "length"),
+    [
+        (0b00000010, 0b00000010, 2),
+        (0b00000010, 0b00010010, 3),
+        (0b00000010, 0b00100010, 4),
+        (0b00000010, 0b00110010, 6),
+        (0b00000010, 0b01000010, 8),
+        (0b00000010, 0b01010010, 12),
+        (0b00000010, 0b01100010, 16),
+        (0b00000010, 0b01110010, 24),
+        (0b01010010, 0b01100010, 16),
+        (0b01110010, 0b00000010, 2),
+        (0b01110010, 0b01000010, 8),
+        (0b11011010, 0b11101010, 16),
+        (0b11110111, 0b11000111, 8),
+        (0b11111110, 0b10001110, 2),
+    ],
+)
+def test_set_preamble_length_bytes(transceiver, mdmcfg1_before, mdmcfg1_after, length):
+    transceiver._spi.xfer.return_value = [15, 15]
+    with unittest.mock.patch.object(
+        transceiver, "_read_single_byte", return_value=mdmcfg1_before
+    ):
+        transceiver.set_preamble_length_bytes(length)
+    transceiver._spi.xfer.assert_called_once_with([0x13 | 0x40, mdmcfg1_after])
+
+
 def test_get_packet_length_bytes(transceiver):
     xfer_mock = transceiver._spi.xfer
     xfer_mock.return_value = [0, 8]